Moved to eventing and made all other resources wait for this collector
